+++ OpenMovement.AxLE.Comms/Values/GoalConfig.cs
@@ -8,6 +8,6 @@ namespace OpenMovement.AxLE.Comms.Values
 
         public UInt32 GoalPeriodOffset { get; set; }
 
-        public UInt16 GoalThreshold { get; set; }
+        public UInt32 GoalThreshold { get; set; }
     }
 }
